Поправил модель
