+++ CollectionAppWithArchitecture/CollectionAppWithArchitecture/Models/Image.swift
@@ -11,36 +11,6 @@ struct Image {
     
     let author: String
     let filename: String
-    
-    static func getImages() -> [Image] {
-        let images: Dictionary<String, String> = [
-            "suganth-l2mIRNeGWoA-unsplash.jpg" : "Suganth",
-            "sincerely-media-FY6-bWUun9Q-unsplash.jpg" : "Sincerely media",
-            "sarah-swinton-_yB6kWwSDAs-unsplash.jpg" : "Sarah Swinton",
-            "priscilla-du-preez-xQMBSGnoSn4-unsplash.jpg" : "Priscilla du Preez",
-            "nicolas-ladino-silva-8IrLhNdHh2w-unsplash.jpg" : "Nicolas Ladino Silva",
-            "nathan-dumlao-71u2fOofI-U-unsplash.jpg" : "Nathan Dumlao",
-            "natanja-grun-X2s8GhnQmds-unsplash.jpg" : "Natanja Grun",
-            "mgg-vitchakorn-frJqgIoK61g-unsplash.jpg" : "Mgg Vitchakorn",
-            "lefteris-kallergis-QsmdVT5pTMw-unsplash.jpg" : "Lefteris Kallergis",
-            "joshua-rawson-harris-bMMOBVp6KQU-unsplash.jpg" : "Joshua Rawson Harris",
-            "jayson-hinrichsen-qRwfk661DnE-unsplash.jpg" : "Jayson Hinrichsen",
-            "jack-dong-2DDGgrJPIjk-unsplash.jpg" : "Jack Dong",
-            "hrihorii-sheldunov-9_H3ss1ko4c-unsplash.jpg" : "Hrihorii Sheldunov",
-            "elsa-noblet-5KD5PmZEfcg-unsplash.jpg" : "Elsa Noblet",
-            "danka-peter-tvicgTdh7Fg-unsplash.jpg" : "Danka Peter",
-            "daniel-korpai-GFrn9Wxqfjo-unsplash.jpg" : "Daniel Korpai",
-            "chuttersnap-rLm4Wq96h_0-unsplash.jpg" : "Chuttersnap",
-            "chris-ralston-Co0-St5PmGY-unsplash.jpg" : "Chris Ralston",
-            "augustine-wong-T0BYurbDK_M-unsplash.jpg" : "Augustine Wong",
-            "alexander-jawfox--upHwnSlzgU-unsplash.jpg" : "Alexander Jawfox",
-            "alexa-suter-Xe0Nj5L5HDw-unsplash.jpg" : "Alexa Suter",
-            "aidan-hancock-OuxSTrA1Z6M-unsplash.jpg" : "Aidan Hancock",
-        ]
-        
-        return images.map { (key, value) in
-            Image(author: value, filename: key)
-        }
-    }
+    let id: UUID
     
 }
